+++ example/windows/flutter/generated_plugin_registrant.cc
@@ -1,0 +1,17 @@
+//
+//  Generated file. Do not edit.
+//
+
+// clang-format off
+
+#include "generated_plugin_registrant.h"
+
+#include <app_badge_plus/app_badge_plus_plugin_c_api.h>
+#include <permission_handler_windows/permission_handler_windows_plugin.h>
+
+void RegisterPlugins(flutter::PluginRegistry* registry) {
+  AppBadgePlusPluginCApiRegisterWithRegistrar(
+      registry->GetRegistrarForPlugin("AppBadgePlusPluginCApi"));
+  PermissionHandlerWindowsPluginRegisterWithRegistrar(
+      registry->GetRegistrarForPlugin("PermissionHandlerWindowsPlugin"));
+}